refactor(event): tech_stack 조회를 MemberClient에서 AdminClient로 교체
