+++ event/src/main/java/com/devticket/event/application/EventService.java
@@ -16,6 +16,7 @@ import com.devticket.event.domain.exception.StockDeductionException;
 import com.devticket.event.domain.model.Event;
 import com.devticket.event.domain.model.EventImage;
 import com.devticket.event.domain.model.EventTechStack;
+import com.devticket.event.infrastructure.client.AdminClient;
 import com.devticket.event.infrastructure.client.MemberClient;
 import com.devticket.event.infrastructure.client.OpenAiEmbeddingClient;
 import com.devticket.event.infrastructure.client.dto.TechStackItem;
@@ -61,6 +62,7 @@ public class EventService {
 
     private final EventRepository eventRepository;
     private final MemberClient memberClient;
+    private final AdminClient adminClient;
     private final ElasticsearchOperations elasticsearchOperations;
     private final ElasticsearchClient esClient;
     private final OpenAiEmbeddingClient openAiEmbeddingClient;
@@ -122,7 +124,7 @@ public class EventService {
     }
 
     private Map<Long, String> buildTechStackMap() {
-        return memberClient.getTechStacks().stream()
+        return adminClient.getTechStacks().stream()
             .collect(Collectors.toMap(TechStackItem::techStackId, TechStackItem::name));
     }
