+++ event/src/main/resources/application-local.yml
@@ -35,6 +35,8 @@ openai:
 service:
   member:
     url: http://localhost:8081
+  admin:
+    url: http://localhost:8087
   ai:
     url: http://localhost:8088
